Initializes and keeps the device token in a Singleton
