+++ app/src/main/AndroidManifest.xml
@@ -143,7 +143,7 @@
         </receiver>
 
         <service android:name=".GcmIntentService" />
-        
+
         <meta-data android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
     </application>
