+++ app/src/main/java/ar/uba/fi/tdp2/trips/Common/BackendService.java
@@ -178,7 +178,7 @@ public interface BackendService {
 
     public static final Retrofit retrofit = new Retrofit.Builder()
 //            .baseUrl("https://private-0e956b-trips5.apiary-mock.com")
-            .baseUrl("http://192.168.1.5")
+            .baseUrl("http://192.168.0.18")
             .addConverterFactory(GsonConverterFactory.create())
             .client(okHttpClient)
             .build();
